+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Fixed message highlight animation alignment in `MessageListView`
 - [[#491]](https://github.com/GetStream/stream-chat-flutter/issues/491): Fix `MediaListView` showing media in wrong order.
+- Fixed `MessageListView` initialIndex not working in some cases.
+- Improved `MessageListView` rendering in case of reordering.
 - Fix image thumbnail generation when using Stream CDN
 
 ✅ Added